// apps/desktop/src/shared/api/generated/BootstrapDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BootstrapDto = {
  onboarding_disposition: string;
  active_game_installation_id: string | null;
  active_profile_id: string | null;
  recovery_summary: string | null;
  app_version: string;
};